Extract the Authentication Helper to API authentication and Authentication
