+++ app/api/bikes/queries.rb
@@ -2,7 +2,7 @@
 
 module Bikes
   class Queries < Grape::API
-    helpers ::Helpers::AuthenticationHelper
+    helpers ::Helpers::ApiAuthenticationHelper
     helpers ::Helpers::OperationAdapter
     before { authenticate! }
