+++ app/api/bikes/reservations.rb
@@ -2,7 +2,7 @@
 
 module Bikes
   class Reservations < Grape::API
-    helpers ::Helpers::AuthenticationHelper
+    helpers ::Helpers::ApiAuthenticationHelper
     helpers ::Helpers::OperationAdapter
     before { authenticate! }
